Removed season functions from AppleTree class and moved TreeInterface to individual file.
