+++ program.php
@@ -3,11 +3,32 @@ namespace Learning;
 
 require_once 'tree.php';
 
-$appleTree = new AppleTree;
+class Season {
+    function __construct(){
+        
+    }
+    function spring(){
+        echo "It's springtime!  ";
+        
+        echo "<br/>";
+	}
+    function summer(){
+        echo "It's summertime!  ";
+        
+        echo"<br/>";
+	}
+    function fall(){
+        echo "It's fall, time to harvest!  ";
+        
+        echo "<br/>";
+    }
+    function winter(){
+        echo "It's wintertime!  ";
+        
+        echo "<br/><br/>";
+    }
+}
 
 for ($i = 0; $i < 10; $i++){
-    $appleTree->spring();
-    $appleTree->summer();
-    $appleTree->fall();
-    $appleTree->winter();
+
 }
