+++ tree.php
@@ -1,17 +1,6 @@
 <?php
 namespace Learning;
-
-interface TreeInterface {
-    function grow();
-    function beginGrowth();
-    function watch();
-    function harvest();
-    function prune();
-    function spring();
-    function summer();
-    function fall();
-    function winter();
-}
+require_once 'tree_interface.php';
 
 class AppleTree implements TreeInterface {
     private $age = 4;
@@ -24,7 +13,7 @@ class AppleTree implements TreeInterface {
         $this->branches = $this->branches + 2;
         echo "The apple tree is $this->age years old, $this->height inches tall, and has $this->branches branches.  ";
 	}
-    function beginGrowth(){
+    function bloom(){
         echo "The apple tree is blooming!  ";
         $this->blossom = ($this->age * 2) * $this->branches;
         echo "It has " . $this->blossom . " blossoms.  ";
@@ -47,25 +36,5 @@ class AppleTree implements TreeInterface {
             echo "Pruned an undesirable branch off the tree.  ";
         }
     }
-    function spring(){
-        echo "It's springtime!  ";
-        $this->grow();
-        $this->beginGrowth();
-        echo "<br/>";
-	}
-    function summer(){
-        echo "It's summertime!  ";
-        $this->watch();
-        echo"<br/>";
-	}
-    function fall(){
-        echo "It's fall, time to harvest!  ";
-        $this->harvest();
-        echo "<br/>";
-    }
-    function winter(){
-        echo "It's wintertime!  ";
-        $this->prune();
-        echo "<br/><br/>";
-    }
+    
 }
