+++ tree_interface.php
@@ -1,0 +1,10 @@
+<?php
+namespace Learning;
+
+interface TreeInterface {
+    function grow();
+    function bloom();
+    function watch();
+    function harvest();
+    function prune();
+}
